+++ project-base/src/Shopsys/ShopBundle/ShopsysShopBundle.php
@@ -5,6 +5,7 @@ namespace Shopsys\ShopBundle;
 use Shopsys\ShopBundle\Component\Translation\Translator;
 use Shopsys\ShopBundle\DependencyInjection\Compiler\RegisterCronModulesCompilerPass;
 use Shopsys\ShopBundle\DependencyInjection\Compiler\RegisterPluginCrudExtensionsCompilerPass;
+use Shopsys\ShopBundle\DependencyInjection\Compiler\RegisterPluginDataFixturesCompilerPass;
 use Shopsys\ShopBundle\DependencyInjection\Compiler\RegisterProductFeedConfigsCompilerPass;
 use Shopsys\ShopBundle\Model\Security\Filesystem\FilemanagerAccess;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -21,6 +22,7 @@ class ShopsysShopBundle extends Bundle
 
         $container->addCompilerPass(new RegisterCronModulesCompilerPass());
         $container->addCompilerPass(new RegisterPluginCrudExtensionsCompilerPass());
+        $container->addCompilerPass(new RegisterPluginDataFixturesCompilerPass());
         $container->addCompilerPass(new RegisterProductFeedConfigsCompilerPass());
     }